Defer initial report
If adding several metrics at once, on the same interval, only the
first will be reported in the initial report due to the immediate
call to `reporter._reportMetricsWithInterval(...)`. By wrapping
that in a `setImmediate(...)`, any other metrics set in that tick
should now also be included in the initial report.