Switch to console-log-level (#62)
* Switch to console-log-level

Bunyan has an optional dependency on dtrace-provider, which is a
native module. With optional dependencies, it's fine for them to
fail to install. On Windows, however, the build process uses
node-gyp and therefore fails the install with a big scary error.

To avoid the build error, I've switched to a simple console logger.

* Add log prefix